Fixed issues with log and pid files

+++ plumbago.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import logging
 import signal
 import yaml
@@ -10,17 +12,21 @@ __author__ = 'uzix'
 
 log = logging.getLogger()
 
+
 class colors:
     GREEN = '\033[92m'
     RED = '\033[91m'
     GRAY = '\033[90m'
+    YELLOW = '\033[93m'
     DEF = '\033[0m'
 
     def disable(self):
         self.GREEN = ''
         self.RED = ''
         self.GRAY = ''
+        self.YELLOW = ''
         self.DEF = ''
+
 
 def createDaemon():
     try:
@@ -47,7 +53,7 @@ def getPlumbagoPid(pidFile):
     try:
         return int(open(pidFile, 'r').readline())
     except:
-        print "Could not open Plumbago pidfile", pidFileOpt
+        print "Could not open Plumbago pidfile", pidFile
         exit(1)
 
 
@@ -61,6 +67,8 @@ def getAlertStatus(alertName):
                 print alert['name'] + ': ' + colors.GREEN + alert['status'] + colors.DEF
             elif alert['status'] == 'DISABLED':
                 print alert['name'] + ': ' + colors.GRAY + alert['status'] + colors.DEF
+            elif alert['status'] == 'UNKNOWN':
+                print alert['name'] + ': ' + colors.YELLOW + alert['status'] + colors.DEF
             else:
                 print alert['name'] + ': ' + colors.RED + alert['status'] + colors.DEF
     elif alertName == 'error':
@@ -71,6 +79,10 @@ def getAlertStatus(alertName):
         for alert in alerts:
             if alert['status'] == 'DISABLED':
                 print alert['name'] + ': ' + colors.GRAY + alert['status'] + colors.DEF
+    elif alertName == 'unknown':
+        for alert in alerts:
+            if alert['status'] == 'UNKNOWN':
+                print alert['name'] + ': ' + colors.YELLOW + alert['status'] + colors.DEF
     else:
         found = False
         for alert in alerts:
@@ -84,19 +96,22 @@ def getAlertStatus(alertName):
                     print 'Status: ' + colors.GREEN + alert['status'] + colors.DEF
                 elif alert['status'] == 'ERROR':
                     print 'Status: ' + colors.RED + alert['status'] + colors.DEF
-                else:
+                elif alert['status'] == 'DISABLED':
                     print 'Status: ' + colors.GRAY + alert['status'] + colors.DEF
+                else:
+                    print 'Status: ' + colors.YELLOW + alert['status'] + colors.DEF
                 break
         if not found:
-            print "\nNo alert exists with that name. Try -s all to see the complete list of alerts\n"
+            print "\nNo alert exists with that name. Try -t all to see the complete list of alerts\n"
     os.remove('/tmp/plumbago.status')
 
 
 def enableAlert(alertName, config, configFile):
     pidFile = config['config']['pidfile']
+    logFile = config['config']['logging']['file']
 
     try:
-        config=yaml.load(open(configFile,'r'))
+        config = yaml.load(open(configFile, 'r'))
     except:
         print 'Could not open config file', configFile
         return
@@ -104,41 +119,39 @@ def enableAlert(alertName, config, configFile):
     if alertName == 'all':
         for alert in config['alerts']:
             try:
-                config['alerts'][alert]['enabled']
+                if not config['alerts'][alert]['enabled']:
+                    config['alerts'][alert]['enabled'] = True
             except:
                 continue
-            if config['alerts'][alert]['enabled'] == False:
-                config['alerts'][alert]['enabled'] = True
-        yaml.dump(config, open(configFile, 'w'))
-        reloadConfig(pidFile)
-        return
-
-    try:
-        config['alerts'][alertName]
-    except:
-        print "\nNo alert exists with that name. Try -s all to see the complete list of alerts\n"
-        return
-
-    try:
-        config['alerts'][alertName]['enabled']
-    except:
-        print "Alert is already enabled"
-        return
-
-    if config['alerts'][alertName]['enabled'] == True:
-        print "Alert is already enabled"
-        return
     else:
-        config['alerts'][alertName]['enabled'] = True
-        yaml.dump(config, open(configFile, 'w'))
-        reloadConfig(config['config']['pidfile'])
+        try:
+            config['alerts'][alertName]
+        except:
+            print "\nNo alert exists with that name. Try -t all to see the complete list of alerts\n"
+            return
+
+        try:
+            if config['alerts'][alertName]['enabled']:
+                print "Alert is already enabled"
+                return
+            else:
+                config['alerts'][alertName]['enabled'] = True
+        except:
+            print "Alert is already enabled"
+            return
+
+    yaml.dump(config, open(configFile, 'w'))
+    config['config']['pidfile'] = pidFile
+    config['config']['logging']['file'] = logFile
+    reloadConfig(config['config']['pidfile'])
 
 
 def disableAlert(alertName, config, configFile):
     pidFile = config['config']['pidfile']
+    logFile = config['config']['logging']['file']
 
     try:
-        config=yaml.load(open(configFile,'r'))
+        config = yaml.load(open(configFile, 'r'))
     except:
         print 'Could not open config file', configFile
         return
@@ -146,19 +159,17 @@ def disableAlert(alertName, config, configFile):
     if alertName == 'all':
         for alert in config['alerts']:
             config['alerts'][alert]['enabled'] = False
-        yaml.dump(config, open(configFile, 'w'))
-        reloadConfig(config['config']['pidfile'])
-        return
+    else:
+        try:
+            config['alerts'][alertName]['enabled'] = False
+        except:
+            print "\nNo alert exists with that name. Try -t all to see the complete list of alerts\n"
+            return
 
-    try:
-        config['alerts'][alertName]
-    except:
-        print "\nNo alert exists with that name. Try -s all to see the complete list of alerts\n"
-        return
-
-    config['alerts'][alertName]['enabled'] = False
     yaml.dump(config, open(configFile, 'w'))
-    reloadConfig(pidFile)
+    config['config']['pidfile'] = pidFile
+    config['config']['logging']['file'] = logFile
+    reloadConfig(config['config']['pidfile'])
 
 
 def reloadConfig(pidFile):
@@ -180,10 +191,6 @@ def terminateServer(pidFile):
 
 
 def startServer(config, configFileOpt):
-    if os.path.exists(config['config']['pidfile']):
-        print "Pid file exists... Maybe plumbago is already running?"
-        exit(0)
-
     print "Starting server..."
     createDaemon()
 
@@ -195,11 +202,7 @@ def startServer(config, configFileOpt):
 
     def handler(signum, frame):
         if signum == signal.SIGUSR1:
-            pidFile = config['config']['pidfile']
-            logFile = config['config']['logging']['file']
             config = yaml.load(open(configFileOpt, 'r'))
-            config['config']['pidfile'] = pidFile
-            config['config']['logging']['file'] = logFile
             server.configure(config)
         elif signum == signal.SIGUSR2:
             server.dump_status()
@@ -213,6 +216,7 @@ def startServer(config, configFileOpt):
 
     server.run()
     os.remove(config['config']['pidfile'])
+    return
 
 
 def definePidFile(config):
@@ -232,6 +236,8 @@ def defineLogFile(config):
 def main():
     from optparse import OptionParser
 
+    server_running = False
+
     parser = OptionParser()
     parser.add_option("-p", "--pid-file", dest="pid", help="Plumbago pid file", metavar="PID_FILE")
     parser.add_option("-c", "--config-file", dest="config", default='./config.yaml', help="Plumbago config file",
@@ -245,8 +251,8 @@ def main():
     parser.add_option("-k", "--kill", dest="kill", action="store_true", default=False, help="Kill Plumbago server")
     parser.add_option("-r", "--reload", dest="reload", action="store_true", default=False,
                       help="Reload Plumbago configuration")
-    parser.add_option("-t", "--status", dest="status", help="Show alerts statuses [alert_name|all|error|disabled]",
-                      metavar="ALERT_NAME")
+    parser.add_option("-t", "--status", dest="status",
+                      help="Show alerts statuses [alert_name|all|error|disabled|unknown]", metavar="ALERT_NAME")
     (options, args) = parser.parse_args()
 
     if not options.reload and not options.status and not options.kill and not options.enable and not options.disable and not options.server:
@@ -260,7 +266,7 @@ def main():
         print "Could not load configuration file", options.config
         exit(1)
 
-    if not options.pid:
+    if not options.pid :
         config['config']['pidfile'] = definePidFile(config)
     else:
         config['config']['pidfile'] = options.pid
@@ -270,33 +276,46 @@ def main():
     else:
         config['config']['logging']['file'] = options.log
 
+    if os.path.exists(config['config']['pidfile']):
+        server_running = True
+
     if options.server:
-        options.reload = False
-        options.status = False
-        options.kill = False
-        options.enable = False
-        options.disable = False
-        startServer(config, options.config)
+        if not server_running:
+            options.reload = False
+            options.status = False
+            options.kill = False
+            options.enable = False
+            options.disable = False
+            startServer(config, options.config)
+        else:
+            print "Plumbago server already running!"
 
-    if options.kill:
-        options.reload = False
-        options.status = False
-        options.enable = False
-        options.disable = False
-        terminateServer(config['config']['pidfile'])
+    if server_running:
+        if options.kill:
+            options.reload = False
+            options.status = False
+            options.enable = False
+            options.disable = False
+            terminateServer(config['config']['pidfile'])
 
-    if options.reload:
-        reloadConfig(config['config']['pidfile'])
+        if options.reload:
+            pidFile = config['config']['pidfile']
+            logFile = config['config']['logging']['file']
+            reloadConfig(config['config']['pidfile'])
+            config['config']['pidfile'] = pidFile
+            config['config']['logging']['file'] = logFile
 
-    if options.status:
-        os.kill(getPlumbagoPid(config['config']['pidfile']), signal.SIGUSR2)
-        getAlertStatus(options.status)
+        if options.status:
+            os.kill(getPlumbagoPid(config['config']['pidfile']), signal.SIGUSR2)
+            getAlertStatus(options.status)
 
-    if options.enable:
-        enableAlert(options.enable, config, options.config)
+        if options.enable:
+            enableAlert(options.enable, config, options.config)
 
-    if options.disable:
-        disableAlert(options.disable, config, options.config)
+        if options.disable:
+            disableAlert(options.disable, config, options.config)
+    elif not options.server:
+        print "Plumbago server not running!"
 
 
 if __name__ == "__main__":
